requirements.txt
albumentations==1.4.21
imageio==2.34.1
nltk==3.8.1
numpy==1.26.4
omegaconf==2.3.0
Pillow==9.1.0
scikit-image==0.22.0
scipy==1.13.1
torch==2.0.1
tqdm==4.64.0
editdistance==0.8.1
